+++ libs/components/src/lib/components/input/input-date-time/input-date-time.component.ts
@@ -48,14 +48,20 @@ import { PRIZM_DATE_RIGHT_BUTTONS } from '../../../tokens/date-extra-buttons';
 import { PrizmDateButton } from '../../../types/date-button';
 import { PRIZM_STRICT_MATCHER } from '../../../constants';
 import { PrizmFormControlHelpers } from '@prizm-ui/helpers';
-import { prizmI18nInitWithKey } from '../../../services';
+import { prizmI18nInitWithKeys } from '../../../services';
 
 @Component({
   selector: `prizm-input-date-time`,
   templateUrl: `./input-date-time.component.html`,
   styleUrls: [`./input-date-time.component.less`],
   changeDetection: ChangeDetectionStrategy.OnPush,
-  providers: [prizmI18nInitWithKey(PRIZM_TIME_TEXTS, 'time'), ...PRIZM_INPUT_DATE_TIME_PROVIDERS],
+  providers: [
+    ...prizmI18nInitWithKeys({
+      time: PRIZM_TIME_TEXTS,
+      dateTexts: PRIZM_DATE_TEXTS,
+    }),
+    ...PRIZM_INPUT_DATE_TIME_PROVIDERS,
+  ],
 })
 export class PrizmInputDateTimeComponent
   extends AbstractPrizmControl<[PrizmDay | null, PrizmTime | null]>