+++ libs/components/src/lib/components/input/input-month-range/input-layout-month-range.component.ts
@@ -35,7 +35,7 @@ import { PRIZM_MONTHS } from '../../../tokens';
   styleUrls: [`./input-layout-month-range.component.less`],
   changeDetection: ChangeDetectionStrategy.OnPush,
   providers: [
-    prizmI18nInitWithKey(PRIZM_MONTHS, 'months'),
+    ...prizmI18nInitWithKey(PRIZM_MONTHS, 'months'),
     PRIZM_MONTH_FORMATTER_PROVIDER,
     PrizmMonthPipe,
     {