+++ libs/components/src/lib/tokens/i18n.ts
@@ -29,35 +29,16 @@ export const PRIZM_MONTHS = new InjectionToken<Observable<PrizmLanguageCore['mon
 
 export const PRIZM_TIME_TEXTS = new InjectionToken<Observable<PrizmLanguageKit['time']>>(`time i18n texts`);
 
-/**
- * @deprecated
- * */
-export const PRIZM_CALENDAR_MONTHS = new InjectionToken(`[TUI_CALENDAR_MONTHS]: short calendar months i18n`, {
-  factory: prizmExtractI18n(`shortCalendarMonths`),
-});
+export const PRIZM_CALENDAR_MONTHS = new InjectionToken<Observable<PrizmLanguageKit['shortCalendarMonths']>>(
+  `[PRIZM_CALENDAR_MONTHS]: short calendar months i18n`
+);
 
-/**
- * @deprecated
- * */
-export const PRIZM_CLOSE_WORD = new InjectionToken(`i18n 'close' word`, {
-  factory: prizmExtractI18n(`close`),
-});
+export const PRIZM_OTHER_DATE_TEXT = new InjectionToken<Observable<PrizmLanguageKit['otherDate']>>(
+  `i18n 'Other date' text`
+);
 
-/**
- * @deprecated
- * */
-export const PRIZM_OTHER_DATE_TEXT = new InjectionToken(`i18n 'Other date' text`, {
-  factory: prizmExtractI18n(`otherDate`),
-});
-
-/**
- * @deprecated
- * */
-export const PRIZM_DATE_TEXTS = new InjectionToken<Observable<Record<PrizmDateMode, string>>>(
-  `date format i18n texts`,
-  {
-    factory: prizmExtractI18n(`dateTexts`),
-  }
+export const PRIZM_DATE_TEXTS = new InjectionToken<Observable<PrizmLanguageKit['dateTexts']>>(
+  `date format i18n texts`
 );
 
 /**